+++ service-rest-api/src/main/java/org/bsc/confluence/rest/RESTConfluenceService.java
@@ -157,6 +157,15 @@ public class RESTConfluenceService extends AbstractRESTConfluenceService impleme
     }
 
     @Override
+    public Model.Page newPage(Model.ID id, String title) {
+        return new Page(Json.createObjectBuilder()
+                        .add( "id", id.toString())
+                        .add( "title", title )
+                        .add( "space", Json.createObjectBuilder().add( "key", ""))
+                        .build());
+    }
+
+    @Override
     public CompletableFuture<Optional<? extends Model.PageSummary>> getPageByTitle(Model.ID parentPageId, String title) {
 
         return childrenPages(String.valueOf(parentPageId))
@@ -276,7 +285,7 @@ public class RESTConfluenceService extends AbstractRESTConfluenceService impleme
     }
 
     @Override
-    public Model.Attachment createAttachment() {
+    public Model.Attachment newAttachment() {
         return new Attachment();
     }